airflow-mcp-rag/pyproject.toml
# pyproject.toml
[project]
name = "airflow-mcp-rag"
version = "0.1.0"
description = "RAG-based MCP server for Apache Airflow documentation"
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.13",
    "langchain-openai>=0.0.5",
    "faiss-cpu>=1.7.4",
    "beautifulsoup4>=4.12.3",
    "requests>=2.31.0",
    "lxml>=5.1.0",
    "mcp>=0.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.3",
    "numpy>=1.26.3",
    "langchain-google-genai>=4.1.3",
    "google-generativeai>=0.8.6",
    "apache-airflow>=2.8.0",
    "duckdb>=0.9.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "black>=24.1.1",
    "ruff>=0.1.0",
]
local = [
    "sentence-transformers>=2.3.1",
]
anthropic = [
    "langchain-anthropic>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data", "src/embeddings", "src/mcp_server", "src/retrieval", "src/utils"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
